requirements.txt
flask
kafka-python
opencv-python
imutils
Pillow